+++ projects/rocprofiler-compute/src/argparser.py
@@ -31,9 +31,9 @@ from typing import Any
 
 
 def print_avail_arch(avail_arch: list[str]) -> str:
-    ret_str = "\t\t\tList all available metrics for analysis on specified arch:"
+    ret_str = "List all available metrics for analysis on specified arch:"
     for arch in avail_arch:
-        ret_str += f"\n\t\t\t   {arch}"
+        ret_str += f"\n   {arch}"
     return ret_str
 
 
@@ -58,6 +58,20 @@ def add_general_group(
     general_group.add_argument(
         "-q", "--quiet", action="store_true", help="Reduce output and run quietly."
     )
+    general_group.add_argument(
+        "--list-metrics",
+        dest="list_metrics",
+        metavar="",
+        choices=supported_archs.keys(),  # ["gfx908", "gfx90a"],
+        help=print_avail_arch(supported_archs.keys()),
+    )
+    general_group.add_argument(
+        "--config-dir",
+        dest="config_dir",
+        metavar="",
+        help="Specify the directory of customized report section configs.",
+        default=rocprof_compute_home.joinpath("rocprof_compute_soc/analysis_configs/"),
+    )
     # Nowhere to load specs from in db mode
     if parser.usage and "database" not in parser.usage:
         general_group.add_argument(
@@ -77,7 +91,9 @@ def omniarg_parser(
 
     ## General Command Line Options
     ## ----------------------------
-    add_general_group(parser, rocprof_compute_version)
+    add_general_group(
+        parser, rocprof_compute_version, supported_archs, rocprof_compute_home
+    )
     parser._positionals.title = "Modes"
     parser._optionals.title = "Help"
 
@@ -111,7 +127,9 @@ Examples:
     )
     profile_parser._optionals.title = "Help"
 
-    add_general_group(profile_parser, rocprof_compute_version)
+    add_general_group(
+        profile_parser, rocprof_compute_version, supported_archs, rocprof_compute_home
+    )
     profile_group = profile_parser.add_argument_group("Profile Options")
     roofline_group = profile_parser.add_argument_group("Standalone Roofline Options")
 
@@ -199,6 +217,12 @@ Examples:
         raise argparse.ArgumentTypeError(f"Invalid metric id: {value}")
 
     profile_group.add_argument(
+        "--list-available-metrics",
+        dest="list_available_metrics",
+        help="\t\t\tList all available metrics for analysis on current arch",
+        action="store_true",
+    )
+    profile_group.add_argument(
         "-b",
         "--block",
         type=validate_block,
@@ -233,13 +257,6 @@ Examples:
         dest="set_selected",
         help="\t\t\tProfile a set of metrics of topic of interest by collecting "
         "counters in a single pass.\n\t\t\tFor available sets, see --list-sets",
-    )
-    profile_group.add_argument(
-        "--config-dir",
-        dest="config_dir",
-        metavar="",
-        help="\t\t\tSpecify the directory of customized report section configs.",
-        default=rocprof_compute_home.joinpath("rocprof_compute_soc/analysis_configs/"),
     )
     profile_group.add_argument(
         "--join-type",
@@ -461,7 +478,9 @@ Examples:
     )
     db_parser._optionals.title = "Help"
 
-    add_general_group(db_parser, rocprof_compute_version)
+    add_general_group(
+        db_parser, rocprof_compute_version, supported_archs, rocprof_compute_home
+    )
     interaction_group = db_parser.add_argument_group("Interaction Type")
     connection_group = db_parser.add_argument_group("Connection Options")
 
@@ -560,7 +579,9 @@ Examples:
     )
     analyze_parser._optionals.title = "Help"
 
-    add_general_group(analyze_parser, rocprof_compute_version)
+    add_general_group(
+        analyze_parser, rocprof_compute_version, supported_archs, rocprof_compute_home
+    )
     analyze_group = analyze_parser.add_argument_group("Analyze Options")
     analyze_advanced_group = analyze_parser.add_argument_group("Advanced Options")
 
@@ -580,10 +601,10 @@ Examples:
         help="\t\tList all detected kernels and kernel dispatches.",
     )
     analyze_group.add_argument(
-        "--list-metrics",
-        metavar="",
-        choices=supported_archs.keys(),  # ["gfx906", "gfx908", "gfx90a"],
-        help=print_avail_arch(supported_archs.keys()),
+        "--list-available-metrics",
+        dest="list_available_metrics",
+        help="\t\tList all available metrics for analysis on current arch",
+        action="store_true",
     )
     analyze_group.add_argument(
         "-k",
@@ -760,13 +781,6 @@ Examples:
         metavar="",
         default=2,
         help="\t\tSpecify desired decimal precision of analysis results. (DEFAULT: 2)",
-    )
-    analyze_advanced_group.add_argument(
-        "--config-dir",
-        dest="config_dir",
-        metavar="",
-        help="\t\tSpecify the directory of customized configs.",
-        default=rocprof_compute_home.joinpath("rocprof_compute_soc/analysis_configs/"),
     )
     analyze_advanced_group.add_argument(
         "--cols",