+++ projects/rocprofiler-compute/src/rocprof_compute_base.py
@@ -85,6 +85,8 @@ class RocProfCompute:
         setattr(self.__args, "loglevel", self.__loglevel)
         set_locale_encoding()
 
+        self.sanitize()
+
         if self.__mode == "profile":
             self.detect_profiler()
         elif self.__mode == "analyze":
@@ -140,6 +142,21 @@ class RocProfCompute:
         else:
             self.__analyze_mode = "cli"
 
+    def sanitize(self):
+        block = False
+        if (hasattr(self.__args, "filter_metrics") and self.__args.filter_metrics) or (
+            hasattr(self.__args, "filter_blocks") and self.__args.filter_blocks
+        ):
+            block = True
+
+        if self.__args.list_metrics is not None and block:
+            console_error("Cannot use --list-metrics with --blocks")
+        if (
+            hasattr(self.__args, "list_available_metrics")
+            and self.__args.list_available_metrics
+        ) and block:
+            console_error("Cannot use --list-available-metrics with --blocks")
+
     @demarcate
     def load_soc_specs(self, sysinfo: Optional[dict] = None) -> None:
         """Load OmniSoC instance for RocProfCompute run"""
@@ -183,6 +200,15 @@ class RocProfCompute:
             if self.__args.specs:
                 print(generate_machine_specs(self.__args))
                 sys.exit(0)
+            elif self.__args.list_metrics is not None:
+                self.list_metrics()
+                sys.exit(0)
+            elif self.__args.config_dir:
+                parser.print_help(sys.stderr)
+                console_error(
+                    "rocprof-compute requires you to pass --list-metrics "
+                    "with --config-dir."
+                )
             parser.print_help(sys.stderr)
             console_error(
                 "rocprof-compute requires you to pass a valid mode. Detected None."
@@ -218,27 +244,40 @@ class RocProfCompute:
 
     @demarcate
     def list_metrics(self) -> None:
+        self.load_soc_specs()
+
+        for_current_arch = False
+        if (
+            hasattr(self.__args, "list_available_metrics")
+            and self.__args.list_available_metrics
+        ):
+            for_current_arch = True
+
         arch = (
-            self.__args.list_metrics
-            if self.__args.list_metrics
-            else self.__mspec.gpu_arch
+            self.__mspec.gpu_arch
+            if (for_current_arch or self.__args.list_metrics is None)
+            else self.__args.list_metrics
         )
-
-        if arch not in self.__supported_archs:
+        if arch in self.__supported_archs.keys():
+            ac = schema.ArchConfig()
+            config_dir = Path(self.__args.config_dir)
+            ac.panel_configs = file_io.load_panel_configs([config_dir.joinpath(arch)])
+            sys_info = (
+                self.__mspec.get_class_members().iloc[0] if for_current_arch else None
+            )
+            parser.build_dfs(archConfigs=ac, filter_metrics=[], sys_info=sys_info)
+            for key, value in ac.metric_list.items():
+                prefix = ""
+                if "." not in str(key):
+                    prefix = ""
+                elif str(key).count(".") == 1:
+                    prefix = "\t"
+                else:
+                    prefix = "\t\t"
+                print(prefix + key, "->", value)
+            sys.exit(0)
+        else:
             console_error("Unsupported arch")
-
-        ac = schema.ArchConfig()
-        ac.panel_configs = file_io.load_panel_configs([
-            str(self.__args.config_dir / arch)
-        ])
-        sys_info = self.__mspec.get_class_members().iloc[0]
-        parser.build_dfs(arch_configs=ac, filter_metrics=[], sys_info=sys_info)
-
-        for key, value in ac.metric_list.items():
-            dot_count = str(key).count(".")
-            prefix = "\t" * dot_count
-            print(f"{prefix}{key} -> {value}")
-        sys.exit(0)
 
     @demarcate
     def list_sets(self) -> None:
@@ -328,7 +367,7 @@ class RocProfCompute:
         self.print_graphic()
         self.load_soc_specs()
 
-        if self.__args.list_metrics is not None:
+        if self.__args.list_metrics is not None or self.__args.list_available_metrics:
             self.list_metrics()
         elif self.__args.list_sets:
             self.list_sets()